mlops-feature-dev_react@660c70449d5/requirements.txt
streamlit==1.52.1
streamlit-option-menu==0.3.6
plotly==5.18.0
flask==3.0.0
flask-cors==4.0.0
orjson==3.11.3
google-cloud-bigquery
google-cloud-aiplatform>=1.66.0
google-auth==2.41.1
pyarrow==16.1.0; python_version < "3.13"
pyarrow>=18.0.0; python_version >= "3.13"
db-dtypes==1.4.3

# Core numerics (py < 3.13)
pandas==2.1.4; python_version < "3.13"
numpy==1.26.2; python_version < "3.13"
scikit-learn==1.3.2; python_version < "3.13"

# Core numerics (py >= 3.13) – versions that publish wheels for 3.13+
pandas==2.2.3; python_version >= "3.13"
numpy==2.1.3; python_version >= "3.13"
# scikit-learn wheels for py3.13 are not available yet; skip to avoid install failures
